app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
azure-cognitiveservices-speech
pydantic
